+++ src/mol-plugin-ui/controls/parameters.tsx
@@ -29,6 +29,7 @@ import { ArrowDownwardSvg, ArrowDropDownSvg, ArrowRightSvg, ArrowUpwardSvg, Book
 import { legendFor } from './legend';
 import { LineGraphComponent } from './line-graph/line-graph-component';
 import { Slider, Slider2 } from './slider';
+import { getColorGradient, getColorGradientBanded } from '../../mol-util/color/utils';
 
 export type ParameterControlsCategoryFilter = string | null | (string | null)[]
 
@@ -666,67 +667,9 @@ export class ColorControl extends SimpleParam<PD.Color> {
     }
 }
 
-function colorEntryToStyle(e: ColorListEntry, includeOffset = false) {
-    if (Array.isArray(e)) {
-        if (includeOffset) return `${Color.toStyle(e[0])} ${(100 * e[1]).toFixed(2)}%`;
-        return Color.toStyle(e[0]);
-    }
-    return Color.toStyle(e);
-}
+const colorGradientInterpolated = memoize1(getColorGradient);
 
-const colorGradientInterpolated = memoize1((colors: ColorListEntry[]) => {
-    if (colors.length === 0) return 'linear-gradient(to right, #000 0%, #000 100%)';
-
-    const hasOffsets = colors.every(c => Array.isArray(c));
-    let styles;
-
-    if (hasOffsets) {
-        const off = [...colors] as [Color, number][];
-        off.sort((a, b) => a[1] - b[1]);
-        styles = off.map(c => colorEntryToStyle(c, true));
-    } else {
-        styles = colors.map(c => colorEntryToStyle(c));
-    }
-
-    return `linear-gradient(to right, ${styles.join(', ')})`;
-});
-
-const colorGradientBanded = memoize1((colors: ColorListEntry[]) => {
-    const n = colors.length;
-    const styles: string[] = [];
-
-    const hasOffsets = colors.every(c => Array.isArray(c));
-    if (hasOffsets) {
-        const off = [...colors] as [Color, number][];
-        // 0 colors present
-        if (!off[0]) {
-            return 'linear-gradient(to right, #000 0%, #000 100%)';
-        }
-        off.sort((a, b) => a[1] - b[1]);
-        styles.push(`${Color.toStyle(off[0][0])} ${(100 * off[0][1]).toFixed(2)}%`);
-        for (let i = 0, il = off.length - 1; i < il; ++i) {
-            const [c0, o0] = off[i];
-            const [c1, o1] = off[i + 1];
-            const o = o0 + (o1 - o0) / 2;
-            styles.push(
-                `${Color.toStyle(c0)} ${(100 * o).toFixed(2)}%`,
-                `${Color.toStyle(c1)} ${(100 * o).toFixed(2)}%`
-            );
-        }
-        styles.push(`${Color.toStyle(off[off.length - 1][0])} ${(100 * off[off.length - 1][1]).toFixed(2)}%`);
-    } else {
-        styles.push(`${colorEntryToStyle(colors[0])} ${100 * (1 / n)}%`);
-        for (let i = 1, il = n - 1; i < il; ++i) {
-            styles.push(
-                `${colorEntryToStyle(colors[i])} ${100 * (i / n)}%`,
-                `${colorEntryToStyle(colors[i])} ${100 * ((i + 1) / n)}%`
-            );
-        }
-        styles.push(`${colorEntryToStyle(colors[n - 1])} ${100 * ((n - 1) / n)}%`);
-    }
-
-    return `linear-gradient(to right, ${styles.join(', ')})`;
-});
+const colorGradientBanded = memoize1(getColorGradientBanded);
 
 function colorStripStyle(list: PD.ColorList['defaultValue'], right = '0'): React.CSSProperties {
     return {
@@ -947,7 +890,9 @@ export class Vec3Control extends React.PureComponent<ParamProps<PD.Vec3>, { isEx
         const p = getPrecision(this.props.param.step || 0.01);
         const value = `[${v[0].toFixed(p)}, ${v[1].toFixed(p)}, ${v[2].toFixed(p)}]`;
         return <>
-            <ControlRow label={label} control={<button onClick={this.toggleExpanded} disabled={this.props.isDisabled}>{value}</button>} />
+            <ControlRow label={label} control={<button onClick={this.toggleExpanded} disabled={this.props.isDisabled}>
+                <div title={value} className='msp-no-overflow' style={{ display: 'block' }}>{value}</div>
+            </button>} />
             {this.state.isExpanded && <div className='msp-control-offset'>
                 <ParameterControls params={this.components} values={v} onChange={this.componentChange} onEnter={this.props.onEnter} />
             </div>}
@@ -1009,7 +954,9 @@ export class Mat4Control extends React.PureComponent<ParamProps<PD.Mat4>, { isEx
         };
         const label = this.props.param.label || camelCaseToWords(this.props.name);
         return <>
-            <ControlRow label={label} control={<button onClick={this.toggleExpanded} disabled={this.props.isDisabled}>{'4\u00D74 Matrix'}</button>} />
+            <ControlRow label={label} control={<button onClick={this.toggleExpanded} disabled={this.props.isDisabled}>
+                <div className='msp-no-overflow' style={{ display: 'block' }}>{'4\u00D74 Matrix'}</div>
+            </button>} />
             {this.state.isExpanded && <div className='msp-control-offset'>
                 {this.grid}
                 <ParameterControls params={this.components} values={v} onChange={this.componentChange} onEnter={this.props.onEnter} />